requirements.txt
torch>=2.0
torchvision>=0.15
opencv-python>=4.8
Pillow>=10.0
flask>=3.0
matplotlib>=3.7
numpy>=1.24
scikit-learn>=1.3
trdg>=1.8